+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/password/RecoverPasswordResponseDto.kt
@@ -3,6 +3,5 @@ package com.app.arcabyolimpo.data.remote.dto.password
 import com.google.gson.annotations.SerializedName
 
 data class RecoverPasswordResponseDto(
-    @SerializedName("message") val message: String
+    @SerializedName("message") val message: String,
 )
-