+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/supplies/GetFiltersDto.kt
@@ -1,0 +1,9 @@
+package com.app.arcabyolimpo.data.remote.dto.supplies
+
+import com.google.gson.annotations.SerializedName
+
+data class GetFiltersDto(
+    @SerializedName("categories") val categories: List<String>? = null,
+    @SerializedName("measures") val measures: List<String>? = null,
+    @SerializedName("workshops") val workshops: List<String>? = null,
+)