+++ app/src/main/java/com/app/arcabyolimpo/data/repository/supplies/SupplyRepositoryImpl.kt
@@ -2,55 +2,61 @@ package com.app.arcabyolimpo.data.repository.supplies
 
 import com.app.arcabyolimpo.data.mapper.supplies.toDomain
 import com.app.arcabyolimpo.data.remote.api.ArcaApi
+import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
+import com.app.arcabyolimpo.domain.model.filter.FilterData
 import com.app.arcabyolimpo.domain.model.supplies.Supply
+import com.app.arcabyolimpo.domain.model.supplies.SupplyBatchExt
 import com.app.arcabyolimpo.domain.repository.supplies.SupplyRepository
 import javax.inject.Inject
+import javax.inject.Singleton
 
 /**
- * Implementation of [SupplyRepository] that manages data retrieval
- * related to supplies from the remote API.
+ * Retrieves detailed information for a specific supply by its [id].
  *
- * This class uses [ArcaApi] to fetch supply data and maps it into
- * domain models to maintain a clean separation between data and domain layers.
+ * This function calls the API to obtain the corresponding [SupplyDto]
+ * and converts it into a [Supply] domain model using the [toDomain] mapper.
+ *
+ * @param id The unique identifier of the supply to retrieve.
+ * @return A [Supply] object containing detailed supply information.
  */
-class SupplyRepositoryImpl
-@Inject
-constructor(
-    private val api: ArcaApi
-) : SupplyRepository {
 
-    /**
-     * Retrieves a list of all available supplies from the API.
-     *
-     * Each item in the response is converted into a [Supply] domain model.
-     * Since this endpoint provides limited data, some fields like
-     * [Supply.unitMeasure] and [Supply.batch] are initialized as empty.
-     *
-     * @return A list of [Supply] objects representing the available supplies.
-     */
-    override suspend fun getSuppliesList(): List<Supply> {
-        val response = api.getSuppliesList()
-        return response.map { dto ->
-            Supply(
-                id = dto.id,
-                name = dto.name,
-                imageUrl = dto.image,
-                unitMeasure = "",
-                batch = emptyList()
-            )
+@Singleton
+class SupplyRepositoryImpl @Inject constructor(
+        private val api: ArcaApi,
+    ) : SupplyRepository {
+        override suspend fun getSuppliesList(): List<Supply> {
+            val response = api.getSuppliesList()
+            return response.map { dto ->
+                Supply(
+                    id = dto.id,
+                    name = dto.name,
+                    imageUrl = dto.image,
+                    unitMeasure = "",
+                    batch = emptyList(),
+                )
+            }
         }
-    }
 
-    /**
-     * Retrieves detailed information for a specific supply by its [id].
-     *
-     * This function calls the API to obtain the corresponding [SupplyDto]
-     * and converts it into a [Supply] domain model using the [toDomain] mapper.
-     *
-     * @param id The unique identifier of the supply to retrieve.
-     * @return A [Supply] object containing detailed supply information.
-     */
-    override suspend fun getSupplyById(id: String): Supply {
-        return api.getSupply(id).toDomain()
+        override suspend fun getSupplyById(id: String): Supply = api.getSupply(id).toDomain()
+
+        override suspend fun filterSupply(filters: FilterDto): List<Supply> {
+            val response = api.filterSupplies(filters)
+            println("DEBUG FILTER RESPONSE: $response")
+            return response.map { dto ->
+                Supply(
+                    id = dto.id,
+                    name = dto.name,
+                    imageUrl = dto.image,
+                    unitMeasure = "",
+                    batch = emptyList(),
+                )
+            }
+        }
+
+        override suspend fun getFilterData(): FilterData = api.getFilterSupplies().toDomain()
+
+        /** -------------------------------------------------------------------------------------- *
+         * getSupplyBatchById -> calls the API to fetch a supply batch by its ID.
+        * --------------------------------------------------------------------------------------- */
+        override suspend fun getSupplyBatchById(id: String): SupplyBatchExt = api.getSupplyBatchById(id).toDomain()
     }
-}