+++ app/src/main/java/com/app/arcabyolimpo/domain/usecase/password/PostPasswordRegistrationUseCase.kt
@@ -8,18 +8,21 @@ import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
 
 class PostPasswordRegistrationUseCase
-@Inject
-constructor(
-    private val repository: PasswordUserRepository,
-) {
-    operator fun invoke(email: String, password: String): Flow<Result<UpdatePassword>> =
-        flow {
-            try {
-                emit(Result.Loading)
-                val response = repository.postUpdatePassword(email, password)
-                emit(Result.Success(response))
-            } catch (e: Exception) {
-                emit(Result.Error(e))
+    @Inject
+    constructor(
+        private val repository: PasswordUserRepository,
+    ) {
+        operator fun invoke(
+            email: String,
+            password: String,
+        ): Flow<Result<UpdatePassword>> =
+            flow {
+                try {
+                    emit(Result.Loading)
+                    val response = repository.postUpdatePassword(email, password)
+                    emit(Result.Success(response))
+                } catch (e: Exception) {
+                    emit(Result.Error(e))
+                }
             }
-        }
-}
+    }